Move batch parser tests into parser folder for clarity
