+++ src/test/java/medistock/inventory/BatchTest.java
@@ -1,7 +1,6 @@
 package medistock.inventory;
 
 import medistock.exception.MediStockException;
-import medistock.parser.Parser;
 import medistock.ui.Ui;
 import org.junit.jupiter.api.Test;
 
@@ -11,43 +10,11 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 public class BatchTest {
-
-    @Test
-    void prepareBatch_missingNameTag_throwsException() {
-        String input = "batch q/10 d/2025-01-01";
-
-        assertThrows(MediStockException.class,
-                () -> Parser.prepareBatch(input));
-    }
-
-    @Test
-    void prepareBatch_incorrectTagOrder_throwsException() {
-        String input = "batch q/10 n/Vyvanse 40mg d/2025-01-01";
-
-        assertThrows(MediStockException.class,
-                () -> Parser.prepareBatch(input));
-    }
-
-    @Test
-    void prepareBatch_quantityNegative_throwsException() {
-        String input = "batch n/Aspirin q/-5 d/2025-01-01";
-
-        assertThrows(MediStockException.class,
-                () -> Parser.prepareBatch(input));
-    }
-
-    @Test
-    void prepareBatch_validFormat_doesNotThrow() {
-        String input = "batch n/Aspirin q/10 d/2025-01-01";
-
-        assertDoesNotThrow(() -> Parser.prepareBatch(input));
-    }
 
     @Test
     void execute_existingItem_increasesQuantity() throws MediStockException {
